+++ CHANGELOG.md
@@ -5,11 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.10.2] / 2025-11-10 - 2026-02-18
+### Features
+- Support changing `VendorId` and `AddInId` in the `.addin` files before start Revit.
 ### Application
 - Support Revit 2027 with framework `net10`.
 - Update `App` to log filter after `IsBusy` response. (Fix: #83)
 ### Console
 - Update `ricaun.Revit.Installation` to version `1.5.1` to change how delete bundle works. (Delete folders or `PackageContents.xml`)
+- Update `ApplicationPluginsDisposable` with `UpdateRevitAddinFile` to update `VendorId` and `AddInId` in the `.addin` files.
+- Update `CreateRevitServer` with `UpdateRevitAddinFileUsingTestMetadata` to update `VendorId` and `AddInId` in the `.addin` files using the test metadata. (Fix: #84)
+	- `ricaun.RevitTest.Application.VendorId` to set the `VendorId` in the `.addin` file.
+	- `ricaun.RevitTest.Application.AddInId` to set the `AddInId` in the `.addin` file. (This should be used carefully, because the `AddInId` should be unique inside Revit.)
 ### Build
 - Update target framework to `net10`.